sizeof data

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	float d;
 	char c;
 
-	printf("size of a char: %zu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(i));
-	printf("size of a long int: %zu byte(s)\n", (unsigned long)sizeof(l));
-	printf("size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(ll));
-	printf("size of a float: %zu byte(s)\n", (unsigned long)sizeof(d));
+	printf("size of a char: %zu byte(s)\n", sizeof(c));
+	printf("Size of an int: %zu byte(s)\n", sizeof(i));
+	printf("size of a long int: %zu byte(s)\n", sizeof(l));
+	printf("size of a long long int: %zu byte(s)\n", sizeof(ll));
+	printf("size of a float: %zu byte(s)\n", sizeof(d));
 	return (0);
 }
